Refactors repo
