+++ .eslintrc
@@ -3,9 +3,10 @@
     "node": true,
     "es6": true
   },
-  "extends": "airbnb",
+  "extends": "airbnb-base",
   "rules": {
     "no-plusplus": 0,
-    "no-use-before-define": 0
+    "no-use-before-define": 0,
+    "arrow-body-style": 0
   }
 }
