+++ .gitignore
@@ -23,9 +23,6 @@ coverage
 # node-waf configuration
 .lock-wscript
 
-# Compiled binary addons (http://nodejs.org/api/addons.html)
-build/Release
-
 # Dependency directories
 node_modules
 jspm_packages
@@ -38,9 +35,6 @@ jspm_packages
 
 # yarn's clean directory
 .yarnclean
-
-# Build directory
-build
 
 # Remote sync
 .remote-sync.json
