--- lib/source-firewall/.eslint
@@ -1,11 +1,0 @@
-{
-  "env": {
-    "node": true,
-    "es6": true
-  },
-  "extends": "airbnb",
-  "rules": {
-    "no-plusplus": 0,
-    "no-use-before-define": 0
-  }
-}
