--- lib/source-firewall/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "source-firewall",
-  "version": "0.0.1",
-  "main": "firewall.js",
-  "repository": "git@github.com:sourcewifi/source-router.git",
-  "author": "Source Co.",
-  "license": "UNLICENSED",
-  "scripts": {
-    "pretest": "yarn lint",
-    "test": "run-p lint tests-only",
-    "tests-only": "mocha test/**/*spec.js",
-    "lint": "eslint ./**/*.{js,jsx}"
-  },
-  "dependencies": {
-    "babel-preset-airbnb": "^2.1.1",
-    "babel-register": "^6.22.0",
-    "bluebird": "^3.4.7",
-    "express": "^4.14.1"
-  },
-  "devDependencies": {
-    "babel-cli": "^6.22.2",
-    "chai": "^3.5.0",
-    "chai-as-promised": "^6.0.0",
-    "eslint": "^3.14.0",
-    "eslint-config-airbnb-base": "^11.0.1",
-    "eslint-plugin-import": "^2.2.0",
-    "mkdirp": "^0.5.1",
-    "mocha": "^3.2.0",
-    "rimraf": "^2.5.4",
-    "yarn-run-all": "^3.1.1"
-  }
-}
