--- lib/source-firewall/test/.eslintrc
@@ -1,10 +1,0 @@
-{
-  "env": {
-    "mocha": true
-  },
-  "rules": {
-    "import/no-extraneous-dependencies": [2, {
-      "devDependencies": true
-    }]
-  }
-}
