+++ test/sanityCheck_spec.js
